requirements.txt
# Core dependencies
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-auth==2.23.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
flake8==6.1.0
black==23.11.0